feat: update tauri dependencies to stable and configure mobile build
- Updates all tauri and tauri-plugin dependencies from beta to the latest stable versions to resolve persistent build failures.
- Configures tauri.conf.json to use a single high-resolution PNG for auto-generating all mobile icons.
- Adds a comprehensive .gitignore file to exclude build artifacts, generated files, and sensitive data from the repository.
- Includes various UI and configuration updates to support the mobile build process.
