+++ src-tauri/Cargo.toml
@@ -21,5 +21,9 @@ tauri-build = { version = "2.3.1", features = [] }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 log = "0.4"
-tauri = { version = "2.7.0", features = [] }
-tauri-plugin-log = "2"
+tauri = { version = "2.7.0", features = ["mobile"] }
+tauri-plugin-log = "2.0.0"
+# Mobile-specific plugins
+tauri-plugin-shell = "2.0.0"
+tauri-plugin-fs = "2.0.0"
+tauri-plugin-dialog = "2.0.0"
