+++ src/hooks/useClock.ts
@@ -1,0 +1,15 @@
+import { useState, useEffect } from 'react';
+
+export const useClock = () => {
+  const [currentTime, setCurrentTime] = useState(new Date());
+
+  useEffect(() => {
+    const timer = setInterval(() => {
+      setCurrentTime(new Date());
+    }, 1000);
+
+    return () => clearInterval(timer);
+  }, []);
+
+  return currentTime;
+};
